+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/di/FactoryModule.kt
@@ -1,5 +1,7 @@
 package com.mercadopago.android.px.internal.di
 
+import com.mercadopago.android.px.internal.datasource.PaymentDataFactory
+import com.mercadopago.android.px.internal.datasource.PaymentResultFactory
 import com.mercadopago.android.px.internal.datasource.TransactionInfoFactory
 
 internal class FactoryModule {
@@ -7,5 +9,22 @@ internal class FactoryModule {
         get() {
             val session = Session.getInstance()
             return TransactionInfoFactory(session.payerPaymentMethodRepository)
+        }
+
+
+    val paymentResultFactory: PaymentResultFactory
+        get() = PaymentResultFactory(paymentDataFactory)
+
+    val paymentDataFactory: PaymentDataFactory
+        get() {
+            val session = Session.getInstance()
+            return PaymentDataFactory(
+                session.discountRepository,
+                session.configurationModule.userSelectionRepository,
+                transactionInfoFactory,
+                session.amountRepository,
+                session.amountConfigurationRepository,
+                session.configurationModule.paymentSettings
+            )
         }
 }